Move Room relation definitions to new package.
